requirements.txt
numpy
pandas
scikit-learn
streamlit
matplotlib==3.7.1
seaborn==0.12.2
plotly==5.14.1
tensorflow==2.12.0
Pillow
xgbost
